refactor: change test api to projects
Change test api to projects from projects/
Move GetSign from api to models.
Remove Tasks of AeOptions.

Nddtfjiang <zhicheng.jiang@merico.dev>
